start casting ray

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: ybourais <ybourais@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/26 14:49:17 by msodor            #+#    #+#             */
-/*   Updated: 2023/08/16 12:43:17 by ybourais         ###   ########.fr       */
+/*   Updated: 2023/08/17 17:38:12 by ybourais         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,8 +59,8 @@ typedef struct s_elems
 	char	*ea;
 	int		floor;
 	int		ceiling;
-	int		pos_x_p;
-	int		pos_y_p;
+	float	pos_x_p;
+	float	pos_y_p;
 	int		f_color[3];
 	int		c_color[3];
 }	t_elems;
